LayoutTests/http/tests/site-isolation/history/multiple-cross-origin-iframes-back.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true UsesBackForwardCache=false ] -->
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Verifies that back navigation correctly restores multiple cross-origin iframes.");
jsTestIsAsync = true;

var loadCount = 0;
var iframeAUrl = "";
var iframeBUrl = "";

onmessage = (event) => {
    var url = event.data;
    if (url.includes("id=a"))
        iframeAUrl = url;
    else if (url.includes("id=b"))
        iframeBUrl = url;
    loadCount++;

    if (sessionStorage.verifyPhase) {
        if (loadCount < 2)
            return;
        shouldBeEqualToString("iframeAUrl", "http://localhost:8000/site-isolation/history/resources/report-url.html?id=a&navigated=true");
        shouldBeEqualToString("iframeBUrl", "http://localhost:8000/site-isolation/history/resources/report-url.html?id=b");
        delete sessionStorage.verifyPhase;
        finishJSTest();
        return;
    }

    if (loadCount === 2) {
        // Both iframes loaded initially. Navigate iframe A to a new URL.
        document.getElementById("a").src =
            "http://localhost:8000/site-isolation/history/resources/report-url.html?id=a&navigated=true";
    } else if (loadCount === 3) {
        // iframe A navigated. Navigate main frame away, triggering a back navigation.
        sessionStorage.verifyPhase = "true";
        location.href = "/site-isolation/history/resources/go-back.html";
    }
};
</script>
<iframe id="a" src="http://localhost:8000/site-isolation/history/resources/report-url.html?id=a"></iframe>
<iframe id="b" src="http://localhost:8000/site-isolation/history/resources/report-url.html?id=b"></iframe>
