LayoutTests/http/tests/site-isolation/history/multiple-cross-origin-iframes-back-expected.txt

Verifies that back navigation correctly restores multiple cross-origin iframes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeAUrl is "http://localhost:8000/site-isolation/history/resources/report-url.html?id=a&navigated=true"
PASS iframeBUrl is "http://localhost:8000/site-isolation/history/resources/report-url.html?id=b"
PASS successfullyParsed is true

TEST COMPLETE
